lima: set native_integers to fix nir optimization
Without .native_integers set, glsl_to_nir will optimize both '<' and
'>=' operators to the same nir_op_slt in some sample shaders.
With this, it is not possible for lima to restore the operator later in
gpir.
No other set of nir optimization flags was able to change that
behaviour.
After setting that flag, nir starts to output nir_op_flt and nir_op_fge
nodes which can be correctly handled by gpir.

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>